useSendType
